refactor: update type from "Int" to "Float" for BitcoinInput.vout and BitcoinInput.sequence
* The value of BitcoinInput.vout and BitcoinInput.sequence from a bitcoin coinbase transaction is currently decoded as 0xffffffff (max value of Uint32), while the "Int" type only represents Int32 value in Scalar

+++ backend/src/modules/bitcoin/input/input.model.ts
@@ -1,4 +1,4 @@
-import { Field, Int, ObjectType } from '@nestjs/graphql';
+import { Field, Float, ObjectType } from '@nestjs/graphql';
 import * as BitcoinApi from 'src/core/bitcoin-api/bitcoin-api.schema';
 import { BitcoinOutput } from '../output/output.model';
 
@@ -7,7 +7,7 @@ export class BitcoinInput {
   @Field(() => String)
   txid: string;
 
-  @Field(() => Int)
+  @Field(() => Float)
   vout: number;
 
   @Field(() => BitcoinOutput, { nullable: true })
@@ -22,7 +22,7 @@ export class BitcoinInput {
   @Field(() => Boolean)
   isCoinbase: boolean;
 
-  @Field(() => Int)
+  @Field(() => Float)
   sequence: number;
 
   public static from(input: BitcoinApi.Input) {
